# openapi.yml
openapi: "3.1.0"
info:
  title: Clawkson API
  description: API for the Clawkson AI assistant platform
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:47821
    description: Local development server
paths:
...
  /api/agents/{id}/container/workspace:
    get:
      summary: List workspace directory contents
      operationId: listWorkspace
      tags: [Containers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - name: path
          in: query
          schema: { type: string, default: "" }
          description: Sub-path within the workspace to list (default is workspace root)
      responses:
        "200":
          description: Directory listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceListing"
        "400":
          description: Path escapes workspace
        "503":
          description: Docker not available
    delete:
      summary: Delete a file or directory from the workspace
      operationId: deleteWorkspaceEntry
      tags: [Containers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceDeleteRequest"
      responses:
        "204":
          description: Deleted
        "400":
          description: Path escapes workspace or is a directory without recursive=true
        "403":
          description: Admin only
        "404":
          description: Path not found
...
components:
  schemas:
...
    WorkspaceEntry:
      type: object
      required: [name, path, is_dir, size]
      properties:
        name:
          type: string
          description: Entry name (file or directory name only)
        path:
          type: string
          description: Path relative to the workspace root
        is_dir:
          type: boolean
        size:
          type: integer
          description: File size in bytes (0 for directories)
        modified_at:
          type: string
          format: date-time
          nullable: true
    WorkspaceListing:
      type: object
      required: [path, entries]
      properties:
        path:
          type: string
          description: The directory that was listed (relative to workspace root)
        entries:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceEntry"
...
    WorkspaceDeleteRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
          description: Path to delete, relative to workspace root
        recursive:
          type: boolean
          default: false
          description: Required when deleting a directory
...